Fixed some bugs, code is breaking after the 4th push to the stack

+++ src/codegen/linux/load_util.rs
@@ -19,7 +19,7 @@ pub fn load_variable(vars: &Vec<Variable>, name: String, datatype: Datatype, byt
     {
         StatementDatatype::Int => {
             bytecode.add_move_mem_to_reg(Register::RBP, &value_pos.to_string(), Register::RAX, SizeType::QWORD);
-            bytecode.add_move_mem_to_reg(Register::RAX, "0", Register::RAX, SizeType::BYTE);
+            bytecode.add_move_mem_to_reg(Register::RAX, "0", Register::RAX, SizeType::QWORD);
         },
         StatementDatatype::Char | StatementDatatype::Bool => {
             bytecode.add_move_mem_to_reg(Register::RBP, &value_pos.to_string(), Register::RAX, SizeType::QWORD);
